+++ lib/sunspot/queue.rb
@@ -7,5 +7,3 @@ end
 
 require "sunspot/queue/version"
 require "sunspot/queue/session_proxy"
-require "sunspot/queue/index_job"
-require "sunspot/queue/removal_job"